// src/service/simulador/dto/penality-chart.dto.ts
export interface PenalityChartDto {
    data: Date;
    contrato: number;
    contratoPonta: number;
    contratoForaPonta: number;
    demandaPonta: number;
    demandaForaPonta: number;
    eustPonta: number;
    eustForaPonta: number;
    addPonta: number;
    addForaPonta: number;
    piuPonta: number;
    piuForaPonta: number;
    pisPonta: number;
    pisForaPonta: number;
}